US-445 [IMP] do not reconnect if the code is executed on a SYNC_SERVER [IMP] relaunch the sync after reconnection [FIX] fix mistakes

+++ bin/updater.py
@@ -449,27 +449,33 @@ def do_upgrade(cr, pool):
 def reconnect_sync_server():
     """Reconnect the connection manager to the SYNC_SERVER if needed"""
 
-    import base64
-    import pooler
-    from tools import config
-
-    # check password file existance
-    credential_filepath = os.path.join(config['root_path'], 'unifield-socket.py')
+    # automatically reconnect if password file exists
+    import tools
+    credential_filepath = os.path.join(tools.config['root_path'], 'unifield-socket.py')
     if os.path.isfile(credential_filepath):
+        import base64
+        import pooler
         f = open(credential_filepath, 'r')
         lines = f.readlines()
+        f.close()
         if lines:
             try:
                 dbname = base64.decodestring(lines[0])
                 password = base64.decodestring(lines[1])
+                logger.info('dbname = %s' % dbname)
                 db, pool = pooler.get_db_and_pool(dbname, pooljobs=False)
-                cr = db.cursor()
-                # delete the credential file
-                os.remove(credential_filepath)
-                # reconnect to SYNC_SERVER
-                connection = pool.get("sync.client.sync_server_connection")
-                connection.connect(cr, 1, password=password)
-            except Exception as e:
-                message = "Impossible to automatically re-connect to the SYNC_SERVER using credentials file."
-                logger.error(message % (password, unicode(e)))
+                # do not execute this code on server side
+                if not pool.get("sync.server.entity"):
+                    cr = db.cursor()
+                    # delete the credential file
+                    os.remove(credential_filepath)
+                    # reconnect to SYNC_SERVER
+                    connection = pool.get("sync.client.sync_server_connection")
+                    connection.connect(cr, 1, password=password)
 
+                    # relaunch the sync
+                    pool.get('sync.client.entity').sync_withbackup(cr, 1)
+                    cr.close()
+            except Exception as e:
+                message = "Impossible to automatically re-connect to the SYNC_SERVER using credentials file : %s"
+                logger.error(message % (unicode(e)))
